+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the "cph-ng" extension will be documented in this file.
 
 ## [Unreleased]
 
-- **feat**: add support for expanding failed test cases in configuration #130
+- **feat**: add support for expanding failed test cases in configuration
+- **feat**: better error stack trace using source map
 
 ## 0.5.0